Fix a SEGV seen with the file from bug 703922 (before any fix for that).
The code encounters an error, does a 'stop' (zstop). It does oparray_cleanup
which does a ref_stack_pop. The problem is that code checks if the "used"
in the current block is less than the "count" to be popped, and if so, it
moves the pointer back by the count, but if the used EQUALS the count, the
pointer is set just above the top of the current stack block.

The case is with the 'dict_stack' which has 1 element in the current block,
and "extension_used" = 14, it exits with the pstack->p at (pstack->bot - 1)
which would be an empty stack, but it hasn't moved to the previous extension
block (doesn't call ref_stack_pop_block). An empty dict_stack is not allowed
for in the interpreter since 'zend' protects against it ordinarily, so we
use the dangling pointer, resulting in a SEGV.

It looks like the same thing could happen to the op_stack if the count being
popped were exactly right so the 'p' pointer ended up just above an extension
block of that stack.