Bug 704089: Improve check for multiple "char maps" in a CMap
For individual code mappings (endnotdefchar/endcidchar) we check if the
previous mapping values are adjacent to the current one, so we can optimize by
merging them into a range. Each map involves 5 entries on the stack, and the
check to trigger the attempt to merge checked for >5 entries on the stack. That
works as long as the number of mappings is large enough that we always have at
least 10 entries (the merging code will cope with unprocessed entries).

In this case, we only have two mappings, when we reach the check for a marge
condidate, there are seven entries on the stack (5 for the current, 2 for the
remaining mapping), and we end up with a stackunderflow error, because code
expects at least 10 objects on the stack.

So, change the check to >9 (rather than >5), before we attempt the merge.

Make the same change for merging ranges.