Implement JPX PassThrough for pdfwrite
Just like JPEG, we want to pass JPX images unchanged to pdfwrite, unless
of course we need to manipulate them.

This shows no diffs on pdfi, because we flate compress the decompressed
image, so it looks identical to the original JPX, though larger. With
Ghostscript, which defaults to DCT encoding the image, it shows marked
quality improvements.