Fix preserving text rendering mode 3
I'm not sure why this didn't cause more of a problem, except that Tr 3
doesn't render anything....

When the text rendering mode was 3 we set up the operation, and then
called the code for Tr 0. However that function added TEXT_DO_DRAW to
the operation, resulting in an invalid combination of flags.

This commit moves the (small) core of pdfi_show_Tr_0 into a new routine
pdfi_show_simple, this allows us to call that routine from both
pdfi_Tr_preserve and pdfi_show_Tr_0, and only set the DO_DRAW flag in
the Tr 0 case (and in _preserve when the Tr mode is not 3).

This shows a single progression with text position and pdfwrite with
pdfi, but I expect more progressions with the Ghostscript integration.