Tidy up the PS integration code
If someone tries to assign a stream to a PDF context which already has
a stream, signal an error.

Clean up the PostScript pdfctx object if we fail to create a context
for the PDF interpreter.

The .PDFInfo operator was checking the wrong object on the stack, which
meant we needed to call it with 2 copies of the pdfctx object on the
operand stack.

If PDFDrawPage gets an error from pdfi_page_render, don't overwrite the
error code.

Rearrange the allocations in zPDFInit and clean up properly if any
of them fail.

Finally, move the allocator 'unwrap' from .PDFClose to the finalize
routine for the pdfctx object.