Rework the PostScript integration for pdfi
This implements all the functions which are documented as being used by
customers, or which I can find evidence have been supplied to customers
as part of a solution, or which I've promised to maintain (eg for
GSView). It also works with the pdf2dsc script and program, provided
the invocation to use the produced DSC file supplies a --permit-file-read
for the PDF file.