pdfwrite - improve PDFA_def.ps for --permit-file-read
The PDFA_def.ps example program needs to read the ICC profile from
disk, in order to embed it in the output PDF file. Since the change to
security this means the ICC profile must be specified on the command
line with an explicit --permit-file-read.

This commit puts the file operation in a stopped context and if it fails
puts up some text explaining the possible reasons, including the use of
--permit-file-read.