Interpolate input data for imagemasks
The PostScript code does this, which means that the old PS-based PDF
interpreter does too.

If we have an imagemask, we are not sending to a high level device, the
interpolate flag is true for the image, and we are scaling in device
space by at least 2, then we interpolate the input data. The output
data may still be interpolated if the scaling is more than 4.

As part of this we need to know if the destination device is a high
level device, so that is now added to the device_setup structure.