pdfwrite - fix case where last character in string is a font switch
For CIDFonts we have to tell that we have switched to a different
descendant font, so that we can create a new descendant.

If we detect a switch (or think we do) on the last character in the
string, then we do not set subfont. The problem is that if we do not
set subfont, we don't reset the local 'str' which contains the accumulated
text so far.

In fact that string has already been flushed, but that takes place far
down the chain updating the text state, and the caller can't tell.

Because we don't reset the string, we don't emit the last character, but
instead emit the same string all over again, leading to corrupted text.

On the other hand, if we exit on the last character *without* breaking,
we don't want to reset subfont (actually ti shouldn't matter, but its
best to be safe).

So, if we reach the end of string condition and subfont is NULL, then
set it.

This condition can't arise with PDF in the existing Ghostscript
PDF interpreter because of the klunky way it handles Width overrides
in CIDFonts, but it could potentially happen with PostScript input, and
due to a different problem (width mismatches in copied and original
fonts), did happen with pdfi.