pdfwrite: replace font directory in FDArray subfonts
pdfwrite replaces the font directory (->dir) entry in copied fonts with its
own font directory in order to handle fonts from interpreters which do not
outlast the device (originally PCL, but also XPS and pdfi) because the
font directory is tied to the life of the interpreter, but copied fonts are
tied to the life of the device.

With pdfi comes with need to handle CIDFontType0 fonts which contain "subfonts"
each of which also has a reference to the font directory.