Font copying - do not compare glyph widths.
The compare_glyphs code checks the widths of the candidate glyphs and
if they are not precisely identical decides the two glyphs differ.

This causes us problems in pdfwrite, where we return the glyph width
including any Widths/W/W2 metrics, not merely the widths from the glyph
description in the font. Even for well behaved fonts there can be a
small difference between the floating point value in the font and the
integer value in the Widths array. For fonts which use the Widths array
to override the glyph metrics this can be even more significant.

This routine is normally called to check if a glyph which exists in a
copied font is the same as the glyph in the current font, which I
believe is generally the case. If that is true then we already go
through all of the tests, the width test is only an early bailout and
one which I believe rarely occurs.

So for the benefit of pdfi we remove the test here, the worst that can
happen is a slight decrease in performance.