Use the 'get_glyph_name' function pointer stored in the context
In order to be able to switch from using the built-in glyph name
functions to the PostScript interpreter ones, when running inside
Ghostscript, we maintain two function pointers in the context which
point to the functions for retrieving glyph names from indices and
indices from names.

In a few places the code was still using the pdfi interpreter function
rather than the function in the pointer.