pdfwrite - fix word and character spacing with xyshow (or equivalent)
Because (I believe) we didn't originally carry the PDF graphics state
members around in the graphics state (we do now), pdfwrite essentially
reverse engineers the graphics state values by looking at the position
of the next piece of text, comparing to the expected position after the
previous piece of text and using the difference to write TJ entries.

It seems that the 'TEXT_REPLACE_WIDTHS' flag, which is set when we do
an xyshow, *potentially* uses a different code path to the regular
text.

The problem is that this code path does not apply either the character
or word spacing to the 'expected' position, which results in the
calculated position being different to the actual position, when there
is word or character spacing. The code deals with this by applying the
difference. Unfortunately, in the next pass (if there is more text)
it then applies the word spacing again, resulting in twice the required
displacement.

We fix this here by applying the character and word spacing to the
redefined character width, just as in the regular case where we apply
these to the 'normal' character width.

This results in no differences in Ghostscript, clearly this path is not
exercised there, but it produces a large number of progressions with
pdfi.