Don't check for 3,1 cmap before retrieving a Unicode code point
Non-symbolic TTF fonts in PDF are supposed to use either a 1,0 or a 3,1 cmap
table, and no others.

We have a PDF with a non-symbolic TTF with ONLY a 0,1 cmap table (Unicode 1.1).

For the vector device callback to retreive the gid, just allow it to drop into
the same code the 3,1 cmap uses, since they are both Unicode variants, it
ought to be close enough to mostly work.