Fix Pattern resource lookup to avoid spurious reference error
We weren't marking and clearing the loop detection when looking up a
Pattern resource. If we were already in a loop detection state (for
example because we are executing a form) and we used the same pattern
twice, then the second and subsequent uses would find the Pattern on
the loop detection stack.

This would cause us to throw an error, and not find the pattern,
resulting in content going missing.