Fix typo in CTM string generation for annotations
There was a missing space between the Tx and Ty members of the string
generated for a CTM array when writing annotations.

Amazingly this only caused problems with 5 files in the test suite, all
of which have rotated pages. This is because the Tx value was being
treated as nearly 0 and the resulting Rect for the annotation was wildly
wrong. For non-rotated pages the value of Tx is zero anyway.