Fix the pdfi_ttf_glyph_name() callback for symbolic fonts
Firstly, I'd incorrectly thought that the character ID was passed to that
callback, when it is, in fact, the GID, so don't look up the GID.

Secondly, if the post table doesn't contain a glyph name, generate a synthetic
one based on the gid value. The glyph name is only required by ps2write.