Set stroke colour = fill colour initially for type 3 glyphs
The result of pdfwrite on some PCL files produces a PDF file containing
type 3 fonts where the BuildChar for the glyphs uses stroke instead of
fill to draw the shape, but does not explicitly set the stroke colour.

Acrobat (and GS) set the stroke and fill colours to be the same, so the
glyph is drawn in the expected colour. This commit (rather laboriously)
does the same thing for type 3 glyphs.