Fix runEPS use of PageCount.
Conceptually, runEPS decides whether or not to sent a showpage at
the end an EPS file dependent on whether a showpage has been sent
by the file itself.

This is currently implemented by looking at the PageCount device
variable. This is bad, because use of Nup can alter this. A better
idea is to look at the current value of ShowPageCount, which
after all is exactly what we want!