Fix bug 704124: Missing Pattern filled area.
The problem was traced to confusion in the overprint and clist logic.
With the pattern-clist, the playback is done with painting operations
as the pattern is tiled. A bitmap pattern is painted with 'copy_planes'
which overprint ignores when it comes from a bitmap pattern, but the
pattern-clist doesn't work the same and performs overprint compositing
as the pattern-clist is played back.

This was caused by the c_overprint_write not writing the overprint
compositor params as needed because 'gs_pattern1_set_color' did not
set the params->is_fill_color.

Fix by setting the is_fill_color param from the pgs->is_fill_color
which causes the c_overprint_write to push the updated values into the
clist.

Besides fixing the customer's file, this also fixes Ad_Quark.ps page 4.