Fix spurious errors when running Pattern PaintProc to fill text
If we draw text, and the text colour is a Pattern, then we end up
running the Pattern PaintProc after we've already executed the BT
operator. This causes all the operations which aren't permitted inside
a BT/ET block to raise warnings.

Obviously these shouldn't raise warnings, so save and restore the text
block count round the execution of the PaintProc to avoid this.