Fix preserved, stroked text rendering modes with a Pattern stroke colour
When preserving text rendering modes, and using a pattern stroke, the
graphics library doesn't load the device colour for the stroke colour,
which means we don't create the pattern tile, and causes an error when
we try to draw the text.

As per the comment; we could modify the graphics library to load the
stroke device colour, as it already does for the fill colour, but that
is potentially wasteful since we generally won't be using it. So here
we explicitly load the stroke device colour if we are preserving text
rendering modes, and the current Ts is a stroking mode.