epo device: Only pass a fillpage on if one was sent.
The N-up device has the effect of supressing all but the first
fillpage for a nest. If the epo-device is spawned under an N-up
device, this means it ends up clearing the complete page with
color 0 (the value to which last_color is initialised).

Here we tweak the device so that it only sends a fillpage if one
has been swallowed.