Some fixes for type 5 halftones
We weren't counting the /Default halftone when calculating the number
of halftone components, leading to an incorrectly sized array.

Because of this we weren't actually storing the /Default halftone at
all, we were ssicking it in a dummy value and discarding it. This led to
a number of problems because the code was assuming it was in index 0.

I hadn't realised that the /Default halftone needs to be stored in the
0'th index of the components and halftone components arrays. This meant
we were storing halftones where the code expected to find the /Default
halftone causing it to do completely the wrong thing.

Since we no longer need a dummy halftone component, lose it.