Added install.sh for gs_jni project
This script installs the required libraries in
/usr/lib so projects using it can be run.

Fixed build scripts on Mac

Added install_darwin.sh and fixed start_darwin.sh allowing Mac
instances of the viewer to run correctFixed build scripts on Mac

Renamed Manifest.txt files to have .md extension

This will be to allow newline characters at en of line, and to match
the internal extension of Java JAR manifest files.

Updated names of build scripts

Done to improve clarity of which scripts are used for which
operating system.