Have configure check for c++17 compatibility
If tesseract/leptonica is present, check if the the compiler is g++ (or
compatible) and if so, check it understands the -std=c++17 option. If not,
drop the OCR devices, and emit a warning message at the end of configure.

If the compiler is not g++ (or compatible) carry on, but emit a warning to
the effect that it's up to the user to ensure the compiler they use can handle
c++17.