GhostPDF - Tidy up headers in ghostpdf.h
This file was originally copied from the XPS interpreter verbatim, and
included a load of .h files which were then included needlessly in
all of the C files.

This removes almost all the .h files from ghostpdf.h; we keep gserrors.h
since so many C files do actually use it to report errors, and
gxgstate.h which is used in ghostpdf.h

The various C and H files have been updated to pull in the include
files they actually need.