Coverity IDs 372257, 372280, 372287, 372293, 372296, 372303
The gs_const_bytestring data type has a uint 'size' member, but the
pdfi_stream_to_buffer can return a buffer limited by an unsigned 64
bit value. Passing a pointer to a 32-bit value and casting as a
pointer to a 64-bit value is bad because it will corrupt memory.
 Also the threshold code can't (obviously) deal with data that large.
so use a temporary variable and if the returned amount is > max_unit
free the string and error out. 3 cases of this.

Off-by-one error calculating the loop bounds check using sizeof
spot_table.

There was a potential (if NumComponents == 0, which should not be
possible, we should at least get the /Default) to try referencing an
unallocated halftone component. Should be impossible but refactor the
code to make certain it is impossible.