Nup: When Nup is in use, force PageUsesTransparency on.
This prevents us suddenly having to flush the nest when
PageUsesTransparency changes. This would otherwuse result
in unexpected changes in the way pages are laid out.

This obviously affects whether we use clist or not in some
cases, but does not (currently at least), I believe, cause
extra work with transparency blending.