Add gxdso_upcall mechanism, and use it in Nup device.
Here, we add a generic mechanism for child devices to send information
to parent devices, and introduce the first use of it.

Using the dev_spec_op mechanism, together with the 'parent' pointers
kept for subclassing, we define a new 'gx_upcall' function that
sends information up to the topmost device (where it can then trickle
down again using the usual dev_spec_op handling).

We define the first such use of this facility as being the
'gx_upcall_device_realloc' event. This should be sent by a device
just before it reallocates the memory it is using for the page
buffer. This allows the Nup device to spot this, and know that it
should flush the nest to avoid rendered pages being lost.