Tweak beginpage/endpage for subclassing.
Our Postscript BeginPage/EndPage handling is done by having
default BeginPage/EndPage routines that call .callbeginpage
and .callendpage.

This means that any Postscript that overrides BeginPage/EndPage
will prevent calls to .callbeginpage and .callendpage being
made. This seems strange to me, but never mind...

These functions find the 'page_device', and then call the
respective page_proc on them.

This means subclassing has no way to intercept and act on
these procs. For example, in a typical situation, I might have
the Nup device wrapping the epo device wrapping the png16m
page device.

zcallbeginpage is called with dev = Nup. get_page_device returns
png16m, and we then call beginpage on that. Nup never gets a
chance to be informed.

So, we need to alter this. There are 3 possible routes I can see
to change this, and it's not clear to me which is best.

1) We could change get_page_device on the Nup device to claim that
Nup is the page device. We would then need to ensure that all parent
devices forward the page calls down to child devices. This should
already be the case for subclassing devices, but I am unsure of the
other potential ramifications for changing the 'page device' away
from being the bottommost device, so not choosing this option.

2) We could get the page_device, then run up the dev->parent chain
to find the topmost device, and call the page_proc on that, relying
on that passing the page calls down to child devices. This works,
but begs the question, why go to that effort when 3 is simpler?

3) We could drop the get_page_device call entirely, and just call
the page_proc on the topmost device, relying on it passing down to
child devices properly.

Here we implement 3. If we ever find a problem with this, we can
reconsider the other options.