pdfwrite - strip trailing NULL from 'special' pdfmark names
Bug #703787 "Extending the catalog /Names entry"

The bug is nothing to do with the actual description :-(

If a name for a Key/Value contains 'non regular' characters then it is
converted to an internal form with two preceding and one terminating
NULLs. I don't know why we do this, I don't really believe it's
required (as noted in gdevpdfm.c line 875). However I'm loath to
alter it without a good reason because there is no practical way to
test this.

When emitting that name, the pdfmark code stripped the two preceding
NULLs, but omitted to strip the terminating NULL, leading to an invalid
string, in this case for the /Dest.

Strip the terminating NULL as well to avoid this.