Nup: Set PageUsesTransparency even when not present.
Several commits ago, I modified the Nup device so that whenever
PageUsesTransparency was specified, we forced it to be true.
This avoided PDF pages flushing the nest as we flip/flop between
pages with and without transparency.

Unfortunately, with parameters that were themselves 'collections'
(like arrays), this would both leak a bit and corrupt them due
to not copying them properly. Here we steal a bit of code from
gsparamx.c to properly recursively copy such things, and hack it
set PageUsesTransparency as part of the copy process.

We also extend that to ensure that it is ALWAYS set (whether it
is specified or not), to ensure that going from (say) a PS job
to a PDF one does not cause the nest to be unnecessarily flushed.

Also, because of the way params report errors, we need to copy
the reported errors back from the copied param list to the original
one. This requires extensions to the param list handling functions.

Because put_params is not always called between the Nup device
being instantiated and the first page started, we need to call
put_params ourselves. To do that write our own install procedure
and get gdevkrnlsclass.c to call that rather than doing the
subclass itself.

Also, incorporate a SEGV fix from Ray.
The devn_put_params can close the device, but if the device is a
child of the Nup device, the Nup device won't have a chance to flush
the nest.

Also, just in case, when we switch a printer device from clist mode
(as we do when we tear down a device), clear the 'pmemdev->base' that
is assumed to be NULL or valid in non-clist mode.