Nup: Tweak page size sensitivity.
Nup will flush the nest whenever it detects the page size has
changed. It does this because the underlying device may reallocate
memory, thus losing the contents rendered so far.

Unfortunately, the test we were using was slightly too sensitive,
causing false positives. We move to a better test here by actually
doing the same calculation that the underlying device does to see
if the width or height in pixels has changed.

This was seen with NestedPageW being 595.08 rather than 595, and
triggering flushes.