pdfwrite - PDF/A - permit non-XMP entries in Docinfo to contain escapes
PDF/A requires that the XMP metadata and the Document information
dictionary entries which have corresponding XMP keys must be the same.
Byte for byte the same. This means that characters in document
information dictionary strings which must be escaped, or represented in
octal, can't be used.

However, 'arbitrary keys', as defined in the pdfmark reference, do not
correlate with keys in the XMP Metadata and do not need to conform to
this restriction.

This commit checks the keys to see if they are XMP keys before checking
to see if they need to be escaped.