Bug 703899: Don't overwrite explicit FontBBox
The FAPI "refined" (replaced!) the FontBBox with one read from the FAPI server.
For example, with TTF it will use the values from the "head" table.

That is required, to get an accurate bounding box for, for example, TTF fonts
or MicroType fonts.

But it can cause problems in certain cases: one is when a Type42 font includes
a "customised" FontBBox for specific purposes (which may not be strictly "legal"
but should work). In a case more in-line with the spec, the values from the
TTF "head" table will be drawn from the complete, original font, which may not
be accurate for a subset font.

Type 42 and Type 1/2 fonts are *always* serialized in the FAPI API, never
handled as a file or full font buffer.

So, this changes FAPI to only do the "refinement" when it's dealing with an
entire font (either in a memory buffer, or direct from a font file).