Consolidate AGL definitions in C
Previously, the only C code that required access to the Adobe Glyph List was
in the vector devices, so we kept the AGL definition with those. Otherwise
it was always in Postscript.

The PDF interpreter in C also requires the AGL, so move the AGL into base, and
give it its own ".dev".

Still to do: Remove the Postscript definition, and create the Postscript
definition(s) in PS VM from the C definition.