GhostPDF - make sure the graphics state is correct for PS operators
We need to make sure we are using the current Ghostscript graphics
state when we assign the file or stream to the PDF interpreter. This is
because we may need to send information (eg DOCINFO) to the device
when we open the file, and if the graphics state is incorrect then the
device may not be correct.