Fix bug 702950. Ignore extra results from a PostScript Type 4 function.
Ghostscrpt requires PostScript calculator function to return the exact number
of results. If not, fn_PtCr_evaluate() stops and throws an error, but
patch_resolve_color_inline() ignores all errors and passes uninitialized
data downstream.

Adobe Acrobat accepts functions that generate the sufficient number of
results and takes the desired arguments from the top of the stack.
The function in the sample file generates one extra argument.

Following Acrobat, allow, but ignore extra results, Propagation of other
errors to the interpreter level is not yet implemented because it requires
numerous changes in the shading code.

Thanks to Peter Cherepanov for this bug report and patch.