GhostPDF - scale and position appearance when preserving
When running an appearance to preserve it for pdfwrite, we still need
to scale and position the CTM before running the Form XObject. If we
don't then we risk the object lying outside the current clip.

The file tests_private/pdf/uploads/Bug697415.pdf with pdfwrite
demonstrates the problem.

To address this I've moved the code to do the scaling and translation
into it's own routine so we can call it both when rendering and
preserving Appearances.