GhostPDF - don't error on 0 point text in FreeText annotations
The file  tests_private/comparefiles/Bug695848.pdf has a FreeText
annotation where the default appearance (DA) does not specify a font
or point size, leading to us rendering it with a 0-point Helvetica.

This was leading to a degenerate matrix and throwing an error. Instead
we now just don't draw the text when the point size is 0.