Bug 704255: Fix bounds checking of fco file name string
When UFST is in use, the build creates a string containing a list of the
accessible FCOs (separated by "gp_file_name_list_separator" characters) which
we loop over, enumerating the fonts available in each.

The bounds checking of that string was just incorrect.

Thanks to Norbert for pointing out the problem, and suggesting part of the
fix.