GhostPDF - Create Lab-compatible Decode array if required
If an image does not have a /Decode array we manufacture one for it.
However, we always created an array with 2 * number of components where
each component was [0 1].

This doesn't work at all well for Lab colour spaces!

Instead we now create a Decode array where the L component is [0 100],
the a component is [mina maxa] and the b component is [minb maxb] as
specified in the PDF Reference. the a and b components minima and maxima
are retrieved from the colour space Range values.