GhostPDF - evaluate ICC colour spaces early for images
The file tests_private/xps/sumatra/2252_-_cmyk_jpeg_image_with_Photoshop_metadata.xps
when converted to PDF with pdfwrite contains an image in an ICCBased
colour space that causes an error when we try to create an ICC link.

However, this does not occur when setting the colour space, or when
setting the colour, this doesn't happen until we set a device colour
which, for images, we never do.

The result of that is that we don't detect a problem with the space
until we have already started rendering it, and all we get then is an
unhelpful 'unknownerror' return.

In this commit, if the image colour space is an ICC space (which it
normally will be) then we set the colour to 0 for every component and
execute gx_set_dev_color(). This will actually try to use the profile.
If that fails then we fall back to trying to create a device space
based on the number of components in the profile.

This resolves the difference with the test file, though it may be worth
investigating why we get an invalid profile.