Change epo device not to unsubclass itself each time.
If you have a device chain A -> epo -> C, then when epo unsubclasses
itself, this moves device C into the memory previously occupied by
epo.

If this happens at 'unexpected' times, callers can be left holding a
pointer to C that is now stale.

The particular instance is running with the Nup device to psdcmyk.
The device chain is therefore Nup -> epo -> psd. We get a put_params
to change the device n parameters (the number of spots). This call
filters down through Nup, epo, and reaches psd. That calls the devn
prn code that spots this change, and closes the device.

The close device jumps up the device chain to close Nup. This decides
to flush the nest of pages rendered so far, so calls the epo device.
This decides to unsubclass itself. This moves psd into the memory
that epo used to occupy, frees the 'psd' pointer, and closes the psd
device.

This exits, returning back up the chain to where the devn prn code
has just called gs_closedevice. That code is holding a dev pointer
pointed to where psd used to be. As soon as it tries to access it
it dies.

In general, it's much less painful for the epo device to avoid
'unpredicated' unsubclassings, so we change it to stay linked in and
just 'disable' itself.