GhostPDF - warn on use of /DeviceN /All
The /All in name is illegal in PDF (it is valid in PostScript). Acrobat
allegedly seems to accept this; if the ink name array only has one
entry and the entry is /All it seems to convert it to a /Sepaaration
/All, which is equivalent.

However sometimes (tests_private/comparefiles/Testform.v1.0.2.pdf) it
throws an error instead. Possibly this is a change in behaviour with
more recent versions of Acrobat.

In any event, we now raise a warning.