GhostPDF - fix some transparency blending problems
The initial fix here is from Michael who has done all the diagnostic
work on this.

The original code was always pushing an Isolated group, if it pushed
a group. This was incorrect.

Michael then pointed out that pdfi_shading_setup_trans()
does a gsave/grestore around pdfi_trans_setup(). But that function
can potentially change the blending mode (and push a group) and if
we grestore after it, then we throw that away.

The reason we do the gsave/grestore is because we must create a path
so that we can give a bbox to the group. If we have a BBox in the
Shading we can just use that. If we don't then we need to use the
current clip. In order to use the current clip we call clippath() to
create a path from it in the graphics state.

Obviously this changes the graphics state, which is something we don't
want to happen, so we gsave/grestore round it. There is similar code in
pdfi_image_setup_trans().

To avoid this, we no longer use the original pdfi_trans_setup() which
was in any case a one line wrapper around the code doing the real work.

Instead we always call the old 'inner' routine, which means we must
always either supply a bbox, or ensure that the current path is valid
and supply NULL for a bbox.

For stroke and fill the current path is fine. For shadings and images
we now gsave, construct a path, get its bbox, grestore and then pass
the bbox to the new pdfi_trans_setup() function (which is the old
'inner' function renamed). This avoids doing a gsave/grestore round
the potential change in blending mode.