pdfi: Add "patching" of font XUID values
The PS PDF interpreter has code to create a custom XUID or patch an existing
XUID, with values derived from hashing name of the input file and the
object number of the font. This was one strategy to deal with having to use the
PS font loading machinery, which works purely on font name.

With pdfi, we no longer rely on loading fonts by name (we use the PDF objects),
but it turns out this is also required by pdfwrite to cope with multiple
different subset CIDFonts which use the same name (including the subset
prefix!!).

This simply does the same XUID patching, but in C rather than Postscript, and
only for CIDFonts.