Fix '/undefined in /--pdfshowpage-finish' seen with Nup device.
In tweaking the epo device, I'd messed up the forwarding of the
'composite' device call. The default subclass handling has some
complex handling to cope with correctly ordering the devices. By
simply forwarding the call on, I was missing this out.

Fix this by calling the default subclass procedure to do the
forwarding.