Fix incorrect type return in dict_param_enumerate.
All the *_param_enumerate functions return a key and a type. All
of them except dict_param_enumerate return the type as being the
type of the key. dict_param_enumerate returns the type of the
next element along.

Looking back 20 or so years, to commit b4896dc33e, we can see that
the element used for the return of key was changed, but not the
element of use for the return of the type.

Update the type return to match.

This should hopefully fix some problems seen in param copying
with Nup.