Bug 703908: Spot malformed tensor patch streams.
When reading tensor patches from a stream as part of a shading,
each patch can be defined anew, or can carry over previous
corners according to the flags in the stream.

We were not checking that our very first patch definition did
not attempt to carry corners over. Spot this and raise an error now.